+++ include/synthizer.h
@@ -474,16 +474,16 @@ struct syz_RouteConfig {
 SYZ_CAPI syz_ErrorCode syz_initRouteConfig(struct syz_RouteConfig *cfg);
 
 SYZ_CAPI syz_ErrorCode syz_routingConfigRoute(syz_Handle context, syz_Handle output, syz_Handle input, const struct syz_RouteConfig *config);
-SYZ_CAPI syz_ErrorCode syz_routingRemoveRoute(syz_Handle context, syz_Handle output, syz_Handle input, float fade_out);
+SYZ_CAPI syz_ErrorCode syz_routingRemoveRoute(syz_Handle context, syz_Handle output, syz_Handle input, double fade_out);
 
 SYZ_CAPI syz_ErrorCode syz_effectReset(syz_Handle effect);
 
 SYZ_CAPI syz_ErrorCode syz_createGlobalEcho(syz_Handle *out, syz_Handle context);
 
 struct syz_EchoTapConfig {
-	float delay;
-	float gain_l;
-	float gain_r;
+	double delay;
+	double gain_l;
+	double gain_r;
 };
 
 SYZ_CAPI syz_ErrorCode syz_globalEchoSetTaps(syz_Handle handle, unsigned int n_taps, const struct syz_EchoTapConfig *taps);
